refactor: BlackJackResult 제거.
