+++ src/main/java/domain/Bet/BetAmount.java
@@ -1,4 +1,4 @@
-package domain.participant;
+package domain.Bet;
 
 import java.util.Objects;
