--- src/main/java/domain/blackjack/BlackJackResult.java
@@ -1,35 +1,0 @@
-package domain.blackjack;
-
-import domain.participant.Participant;
-
-import java.util.LinkedHashMap;
-import java.util.Map;
-import java.util.Set;
-
-public class BlackJackResult {
-
-    private final LinkedHashMap<Participant, WinStatus> resultByParticipant;
-
-    public BlackJackResult(final LinkedHashMap<Participant, WinStatus> resultByParticipant) {
-        this.resultByParticipant = resultByParticipant;
-    }
-
-    public int getDealerWinCount() {
-        long dealerWinCount = resultByParticipant.values().stream()
-                .filter(isWin -> WinStatus.LOSE == isWin)
-                .count();
-        return (int) dealerWinCount;
-    }
-
-    public Set<Map.Entry<Participant, WinStatus>> getEntry() {
-        return resultByParticipant.entrySet();
-    }
-
-    public LinkedHashMap<Participant, WinStatus> getResultByParticipant() {
-        return resultByParticipant;
-    }
-
-    public int getTotalCount() {
-        return resultByParticipant.size();
-    }
-}
